Sleep if timeout is hit during default attachments processing (#40)
## Summary
This pull request introduces a minor version update to the library and
adds a new utility for handling timeouts, which is integrated into the
`WorkerAdapter` class. This mitigates the issue of lambda timing out.
The most important changes include the addition of a `sleep` function
and a default sleep delay constant, as well as their usage in the
`WorkerAdapter` to handle timeouts.

## Related Issues
- work-item: https://app.devrev.ai/devrev/works/ISS-179204

## Type of Change
- [x] Change doesn't affect products or customers
- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] Non-breaking change (the new functionality and code refactor do
not require a migration strategy)
- [ ] Breaking change (fix or feature that will require a migration plan
for data or other services)
- [ ] Documentation/comment update
- [ ] Other (please describe):

## Testing Procedure
- Run any ADaaS snap-in where you experience problems with lambda
timeout-ing during attachments extraction, using this library version:
1.5.1

## Checklist
- [x] I used generative AI to generate this PR
- [x] I have self-reviewed my code for clarity and correctness
- [ ] I have added or updated comments for complex or non-obvious logic
in my code
- [ ] I have updated relevant documentation (e.g., README, code docs)
- [x] My changes do not introduce new warnings or errors
- [ ] I have added or updated tests to cover new or changed
functionality
- [x] All tests pass locally with my changes applied